Refactor: Correct verification token import and update dependencies
Fix typo in verification token import path and update project dependencies to latest versions.
